refactor: icons 파일명 수정에 따른 주소 변경
